+++ fiddles/angular2/fiddle-0018-AgGrid/src/DataJsonApi.ts
@@ -1,0 +1,17 @@
+import {Component, enableProdMode, Injectable} from 'angular2/core';
+import {Http, Headers, HTTP_PROVIDERS} from 'angular2/http';
+import 'rxjs/add/operator/map';
+
+
+
+@Injectable()
+export class DataJsonApi {
+    constructor(private http: Http) {}
+    request() {
+        const endpoint = "data.json";
+        return this.http
+            .get(endpoint)
+            .map(res => res.json());
+    }
+}
+